Разобьет код на модули
